CodeQL #30, #31: widen two more multiplication callsites to size_t
The earlier PoolAlloc(size_t) change in 102107b made CodeQL aware of
two more (unsigned int)new_stride * new_h callsites in img_io.cpp's
chop and crop-to-content paths - same pattern as the five I fixed
last commit, just newly visible now that the destination type is
wider. Same fix: cast one operand to size_t so the multiply lifts to
size_t.

After this commit, every cpp/integer-multiplication-cast-to-long
alert in our own code is resolved. The 20 remaining open alerts are
all in third-party headers (stb_image*, imgui/*) and not ours to fix.

+++ platform/img_io.cpp
@@ -996,7 +996,7 @@ int ChopMarkedImages(int grid_w, int grid_h, bool trim)
                 new_img->w = new_w;
                 new_img->h = new_h;
                 int new_stride = (new_w + 3) & ~3;
-                new_img->data_p = (unsigned char *)PoolAlloc((unsigned int)new_stride * new_h);
+                new_img->data_p = (unsigned char *)PoolAlloc((size_t)new_stride * new_h);
                 if (!new_img->data_p) continue;
                 memset(new_img->data_p, 0, (size_t)new_stride * new_h);
 
@@ -1144,7 +1144,7 @@ int CropMarkedImagesToContent(void)
         int new_w = max_x - min_x + 1;
         int new_h = max_y - min_y + 1;
         int new_stride = (new_w + 3) & ~3;
-        unsigned char *dst = (unsigned char *)PoolAlloc((unsigned int)new_stride * new_h);
+        unsigned char *dst = (unsigned char *)PoolAlloc((size_t)new_stride * new_h);
         if (!dst) continue;
         memset(dst, 0, (size_t)new_stride * new_h);
         for (int y = 0; y < new_h; y++) {
